07_Advanced_classifier/requirements.txt
streamlit
chromadb
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
pydantic
requests
beautifulsoup4
python-dotenv
schedule
pdfplumber
pymupdf
pytesseract
browser-cookie3
playwright